requirements_pip.txt
Django==3.2.9
psycopg2>=2.8
djangorestframework==3.14.0
python-dotenv==0.21.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1